:recycle: Resolve Comments

+++ contracts/Formulas/Formulas.sol
@@ -15,10 +15,10 @@ contract Formulas is IFormulas {
      * @dev Calculate the discount amount:
      * (Discount Fee (%) * (Advanced Amount / 365) * (Finance Tenure - Late Days))
      * @return uint Amount of the Discount
-     * @param discountFee uint24 input from user
-     * @param financeTenure uint16 input from user
-     * @param lateDays uint16 calculated based on user inputs
-     * @param advancedAmount uint calculated based on user inputs
+     * @param discountFee, uint24 input from user
+     * @param financeTenure, uint16 input from user
+     * @param lateDays, uint16 calculated based on user inputs
+     * @param advancedAmount, uint calculated based on user inputs
      */
     function discountAmountCalculation(
         uint discountFee,
@@ -27,74 +27,68 @@ contract Formulas is IFormulas {
         uint advancedAmount
     ) external pure returns (uint) {
         return
-            (((uint24(discountFee) * advancedAmount) *
-                (uint16(financeTenure) - uint16(lateDays))) / 365) / _PRECISION;
+            ((discountFee * advancedAmount * (financeTenure - lateDays)) /
+                365) / _PRECISION;
     }
 
     /**
      * @dev Calculate the advanced amount: (Invoice Limit * Advance Ratio)
      * @return uint Advanced Amount
-     * @param invoiceLimit uint input from user
-     * @param advanceRatio uint16 input from user
+     * @param invoiceLimit, uint input from user
+     * @param advanceRatio, uint16 input from user
      */
     function advancedAmountCalculation(uint invoiceLimit, uint advanceRatio)
         external
         pure
         returns (uint)
     {
-        return (invoiceLimit * uint16(advanceRatio)) / _PRECISION;
+        return (invoiceLimit * advanceRatio) / _PRECISION;
     }
 
     /**
      * @dev Calculate the factoring amount: (Invoice Amount * Factoring Fee)
      * @return uint Factoring Amount
-     * @param invoiceAmount uint input from user
-     * @param factoringFee uint24 input from user
+     * @param invoiceAmount, uint input from user
+     * @param factoringFee, uint24 input from user
      */
     function factoringAmountCalculation(uint invoiceAmount, uint factoringFee)
         external
         pure
         returns (uint)
     {
-        return (invoiceAmount * uint24(factoringFee)) / _PRECISION;
+        return (invoiceAmount * factoringFee) / _PRECISION;
     }
 
     /**
      * @dev Calculate the late amount: (Late Fee (%) * (Advanced Amount / 365) * Late Days)
      * @return uint Late Amount
-     * @param lateFee uint24 input from user
-     * @param lateDays uint16 calculated based on user inputs
-     * @param advancedAmount uint calculated based on user inputs
+     * @param lateFee, uint24 input from user
+     * @param lateDays, uint16 calculated based on user inputs
+     * @param advancedAmount, uint calculated based on user inputs
      */
     function lateAmountCalculation(
         uint lateFee,
         uint lateDays,
         uint advancedAmount
     ) external pure returns (uint) {
-        return
-            ((uint24(lateFee) * (advancedAmount) * uint16(lateDays)) / 365) /
-            _PRECISION;
+        return ((lateFee * advancedAmount * lateDays) / 365) / _PRECISION;
     }
 
     /**
      * @dev Calculate the number of late days: (Payment Receipt Date - Due Date - Grace Period)
      * @notice Number of late days will never be less than ‘0’
      * @return uint16 Number of Late Days
-     * @param paymentReceiptDate uint48 input from user or can be set automatically
-     * @param dueDate uint48 input from user
-     * @param gracePeriod uint16 input from user
+     * @param paymentReceiptDate, uint48 input from user or can be set automatically
+     * @param dueDate, uint48 input from user
+     * @param gracePeriod, uint16 input from user
      */
     function lateDaysCalculation(
         uint paymentReceiptDate,
         uint dueDate,
         uint gracePeriod
     ) external pure returns (uint) {
-        if (uint48(dueDate) <= uint48(paymentReceiptDate))
-            return
-                uint(
-                    ((uint48(paymentReceiptDate) - uint48(dueDate)) / 1 days) -
-                        uint16(gracePeriod)
-                );
+        if (dueDate <= paymentReceiptDate)
+            return ((paymentReceiptDate - dueDate) / 1 days) - gracePeriod;
 
         return 0;
     }
@@ -102,22 +96,22 @@ contract Formulas is IFormulas {
     /**
      * @dev Calculate the invoice tenure: (Due Date - Invoice Date)
      * @return uint16 Invoice Tenure
-     * @param dueDate uint48 input from user
-     * @param invoiceDate uint48 input from user
+     * @param dueDate, uint48 input from user
+     * @param invoiceDate, uint48 input from user
      */
     function invoiceTenureCalculation(uint dueDate, uint invoiceDate)
         external
         pure
         returns (uint)
     {
-        return uint((uint48(dueDate) - uint48(invoiceDate)) / 1 days);
+        return (dueDate - invoiceDate) / 1 days;
     }
 
     /**
      * @dev Calculate the reserve amount: (Invoice Amount - Advanced Amount)
      * @return uint Reserve Amount
-     * @param invoiceAmount uint input from user
-     * @param advancedAmount uint calculated based on user inputs
+     * @param invoiceAmount, uint input from user
+     * @param advancedAmount, uint calculated based on user inputs
      */
     function reserveAmountCalculation(uint invoiceAmount, uint advancedAmount)
         external
@@ -130,28 +124,24 @@ contract Formulas is IFormulas {
     /**
      * @dev Calculate the finance tenure: (Payment Receipt Date - Date of Funds Advanced)
      * @return uint16 Finance Tenure
-     * @param paymentReceiptDate uint48 input from user
-     * @param fundsAdvancedDate uint48 input from user
+     * @param paymentReceiptDate, uint48 input from user
+     * @param fundsAdvancedDate, uint48 input from user
      */
     function financeTenureCalculation(
         uint paymentReceiptDate,
         uint fundsAdvancedDate
     ) external pure returns (uint) {
-        return
-            uint(
-                (uint48(paymentReceiptDate) - uint48(fundsAdvancedDate)) /
-                    1 days
-            );
+        return (paymentReceiptDate - fundsAdvancedDate) / 1 days;
     }
 
     /**
      * @dev Calculate the total fees amount:
      * (Factoring Amount + Discount Amount + Additional Fee + Bank Charges Fee)
      * @return uint Total Amount
-     * @param factoringAmount uint input from user
-     * @param discountAmount uint input from user
-     * @param additionalFee uint input from user
-     * @param bankChargesFee uint input from user
+     * @param factoringAmount, uint input from user
+     * @param discountAmount, uint input from user
+     * @param additionalFee, uint input from user
+     * @param bankChargesFee, uint input from user
      */
     function totalFeesCalculation(
         uint factoringAmount,
@@ -167,9 +157,9 @@ contract Formulas is IFormulas {
      * @dev Calculate the net amount payable to the client:
      * (Total amount received – Advanced amount – Total Fees)
      * @return uint Net Amount Payable to the Client
-     * @param totalAmountReceived uint calculated based on user inputs
-     * @param advancedAmount uint calculated based on user inputs
-     * @param totalFees uint24 calculated based on user inputs
+     * @param totalAmountReceived, uint calculated based on user inputs
+     * @param advancedAmount, uint calculated based on user inputs
+     * @param totalFees, uint24 calculated based on user inputs
      */
     function netAmountPayableToClientCalculation(
         uint totalAmountReceived,
@@ -183,8 +173,8 @@ contract Formulas is IFormulas {
      * @dev Calculate the total amount received:
      * (Amount Received from Buyer + Amount Received from Supplier)
      * @return uint Total Received Amount
-     * @param buyerAmountReceived uint input from user
-     * @param supplierAmountReceived uint input from user
+     * @param buyerAmountReceived, uint input from user
+     * @param supplierAmountReceived, uint input from user
      */
     function totalAmountReceivedCalculation(
         uint buyerAmountReceived,
